feat: Refactor transaction handling and update contract address
Significant changes have been made to the way transactions are handled. The transaction-related code has been removed from AuthorizedPage.tsx and moved to RequestModal.tsx. This change simplifies the structure of AuthorizedPage.tsx, making it easier to understand and maintain.

In addition, the PAYMENT_HANDLER_ADDRESS in contracts.ts has been updated with a new value. This change ensures that our application is interacting with the correct smart contract on the blockchain.

Please note that some functions in SendModal.tsx have been temporarily disabled as part of this refactor. These will be re-enabled in future commits once their implementation is finalized.